Split startup-input funnel out of RestorableAgentSession.swift
Keeps the file inside its length budget: the resume/fork startup-input
methods and AgentResumeScriptStore move to
RestorableAgentSessionStartupInput.swift unchanged.
